jl - passing local tests and all mutations for RecommendationRequest form
